Add initial state of ZoneGraph Plugin (UE5.4.2) to Tempo. (#78)
Now, ZoneGraph will be a "Project Plugin". And, in order to get all the existing Engine Plugin modules to reference this Project Plugin version of ZoneGraph, we need to make a modification to the UnrealBuildTool, using our "InstallToolChain.sh" script.

So, we copy our modified UEBuildTarget.cs file into UnrealBuildTool. This only happens when necessary and has checks to prevent modifying
unexpected versions (ie. future versions) of the file.