Enable turning on/off vehicle headlights with environmental brightness. (#142)
This is done by syncing the Brightness value from an ATempoBrightnessMeter Actor that gets automatically spawned in the level to Mass via UTempoMassAgentBrightnessMeterSyncTrait. The user can place an ATempoBrightnessMeter Actor in the level themselves to override its placement, and the placed ATempoBrightnessMeter Actor will be used instead.

ISM representations will always have their headlights off for now.

Note: This was done in support of enabling vehicle colors and vehicle lights. Therefore, this change also includes a fix to allow Mass Traffic to generate a new seed (to allow randomization between runs by default) as advertised when UMassTrafficSettings::RandomSeed is set to 0.