Add uniform sub-sampling along roads and road modules for better fit ZoneShapes in curvy sections. (#190)
USplineComponent splines are cubic Hermite splines with independent leave and arrive tangents. These will most often be used to define paths for roads and sidewalks and such.

However, the intermediate ZoneGraph representation needs to be described in terms of FZoneShapePoints, which are used to describe cubic Bezier splines.

In general, exactly preserving the represented path is possible when converting between Hermite splines and Bezier splines. However, in this case, the independent leave and arrive tangents of the Hermite spline can't be input to the FZoneShapePoints because they only have one Rotation field and one TangentLength field. So, this method of sub-sampling attempts to follow the original path precisely without taking on the work of modifying the FZoneShapePoint data structure to accommodate the flexibility of independent leave and arrive tangents and all the downstream effects that might cause at this time.

Further, Tempo's ZoneGraph generation is meant to be "source path representation agnostic" and sampling in this manner should produce fairly well-fitting ZoneShape paths for various source path representations.