Add processor for detecting and resolving yield cycle deadlocks.
Each frame, it checks for a yield cycle.  Then, it resolves it by
allowing all the Entities that were yielding at the time on one
particular lane in the yield cycle to ignore their yield logic until
they're able to make it to their next lane.

It will choose the lane with an Entity (or group of Entities), where one
of those Entities has a normalized distance along its lane that is
greater than any Entity on the other lanes in the yield cycle.

It will also choose only among pedestrian lanes, if they are in the
yield cycle.  Otherwise, it will choose among vehicle lanes.  This gives
a preference to allow pedestrians in a yield cycle deadlock to be the
ones to ignore their yield logic, potentially walking through vehicles
until they reach the other side of the crosswalk, thereby resolving the
yield cycle deadlock and allowing the other traffic to flow freely when
the opportunity presents itself.