Remove pre-emptive yield behavior.
It will only cause problems now since it was originally designed to
roll-out slightly, then yield to neighboring vehicles entering the
intersection from the same side.

- During the roll-out, it doesn't look for pedestrians, which will now
  potentially be in the entrance crosswalks at interections at the time
  vehicles are otherwise allowed to enter the intersection.
- It uses the old simple distance-based yield logic for yielding to both
  vehicles and pedestrians.

If a "roll-out" behavior is desired, it should be re-implemented, such
that it integrates with all the latest checks stemming from the reactive
yield logic.