Fix issue with vehicles running-over pedestrians at entrance crosswalks.
Sometimes, vehicles would run over pedestrians at intersection entrance
crosswalks after they "crept" into the crosswalk lanes, while waiting
for an opportunity to merge.  This maintains that the vehicle is in
"distance conflict" with pedestrians even as the vehicle enters the
crosswalk lane, if those pedestrians are in the vehicle's lane.